Fix missing ViewControls and nose geometry memo
